+++ src/DotSama.ts
@@ -1,10 +1,24 @@
 import { ApiPromise, WsProvider } from '@polkadot/api'
 import { KeyringPair } from '@polkadot/keyring/types'
+import { blake2AsHex } from '@polkadot/util-crypto'
+import { decodeAddress, encodeAddress } from '@polkadot/keyring'
+import { u8aToHex, isHex, hexToBn } from '@polkadot/util'
+import { SubmittableExtrinsic } from '@polkadot/api/types'
+import { ISubmittableResult } from '@polkadot/types/types'
+import Buffer from 'buffer'
 import BN from 'bn.js'
+
 import { DEFAULT_TOKEN_DECIMALS } from './constants/chain'
-import toBn from './utils/toBn'
+import { Options } from './types/options'
+import signTx from './utils/signTx'
+import { TCrowdloanFund } from './types/crowdloanFund'
+import { TContribution } from './types/crowdloanContribution'
+import { TCrowdloan } from './types/crowdloan'
+import { fromBN } from './utils/fromBN'
+import { toBN } from './utils/toBN'
 /**
  * @class DotSama
+ * @author Mangata Finance
  * The DotSama class defines the `getInstance` method that lets clients access the unique singleton instance.
  */
 export class DotSama {
@@ -63,14 +77,77 @@ export class DotSama {
     api.disconnect()
   }
 
-  async createCrowdloan() {}
-
-  async getCrowdloanFunds() {}
-
-  async contributeToCrowdloan(account: KeyringPair, parachainId: string, amount: string) {
+  async getCrowdloanFunds(parachainId: string) {
     const api = await this.getApi()
-    const valueToContribute = toBn(amount, this.decimals)
-    const extrinsic = api.tx.crowdloan.contribute(parachainId, valueToContribute, null)
+    const parentHash = (await api.rpc.chain.getHeader()).parentHash
+
+    const fund = (await api.query.crowdloan.funds(parachainId)).toJSON() as TCrowdloanFund
+    let bytes = new TextEncoder().encode('crowdloan')
+    let buf = Buffer.Buffer.allocUnsafe(4)
+    buf.writeUInt32LE(fund.trieIndex)
+    var array = Uint8Array.from(buf)
+    var concatArray = new Uint8Array([...bytes, ...array])
+    let childEncodedBytes = u8aToHex(new TextEncoder().encode(':child_storage:default:'))
+    let crowdloanKey = childEncodedBytes + blake2AsHex(concatArray, 256).substring(2)
+    let allKeys = (await api.rpc.childstate.getKeys(crowdloanKey, null, parentHash)) as any
+
+    let crowdloan: TCrowdloan = {
+      cap: isHex(fund.cap) ? fromBN(hexToBn(fund.cap)) : fromBN(new BN(fund.cap)),
+      raised: isHex(fund.raised) ? fromBN(hexToBn(fund.raised)) : fromBN(new BN(fund.raised)),
+      contributions: [] as TContribution[],
+    }
+
+    const storages = await api.rpc.childstate.getStorageEntries(
+      crowdloanKey,
+      allKeys.map((i: any) => i.toHex())
+    )
+
+    for (let i = 0; i < storages.length; i++) {
+      const storage = storages[i]
+      if (storage.isSome) {
+        let storage_item = storage.unwrap()
+        let balance = BigInt(u8aToHex(storage_item.slice(0, 16).reverse()))
+        let memoLenght = storage.unwrap().slice(16, 17)
+        let memo = ''
+        if (u8aToHex(memoLenght) != '0x00') {
+          const hexAddress = u8aToHex(storage_item.slice(17, storage_item.length))
+          memo = encodeAddress(hexAddress, 42)
+        }
+
+        crowdloan.contributions.push({
+          kusamaAccount: encodeAddress(allKeys[i].toHex(), 2),
+          contribution: fromBN(new BN(balance.toString())),
+          memo: memo,
+        })
+      }
+    }
+
+    return crowdloan
+  }
+
+  async contributeToCrowdloan(
+    account: string | KeyringPair,
+    parachainId: string,
+    amount: string,
+    options?: Options
+  ) {
+    const api = await this.getApi()
+    const valueToContribute = toBN(amount, this.decimals)
+
+    let extrinsic: SubmittableExtrinsic<'promise', ISubmittableResult>
+
+    if (options?.memo) {
+      const memoAddress = decodeAddress(options.memo)
+      const memo = u8aToHex(memoAddress)
+      const valueToContribute = toBN(amount, this.decimals)
+      extrinsic = api.tx.utility.batchAll([
+        api.tx.crowdloan.contribute(parachainId, valueToContribute, null),
+        api.tx.crowdloan.addMemo(parachainId, memo),
+      ])
+    } else {
+      extrinsic = api.tx.crowdloan.contribute(parachainId, valueToContribute, null)
+    }
     // now we need to sign the transaction
+    await signTx(api, account, extrinsic, options)
   }
 }