+++ src/index.ts
@@ -6,3 +6,11 @@ export {
   WESTEND_ENDPOINT,
   ROCOCO_ENDPOINT,
 } from './constants/endpoints'
+
+export { Options } from './types/options'
+export { TCrowdloan } from './types/crowdloan'
+export { TContribution } from './types/crowdloanContribution'
+export { TCrowdloanFund } from './types/crowdloanFund'
+
+export { toBN } from './utils/toBN'
+export { fromBN } from './utils/fromBN'