+++ src/types/crowdloan.ts
@@ -1,0 +1,7 @@
+import { TContribution } from './crowdloanContribution'
+
+export type TCrowdloan = {
+  cap: string
+  raised: string
+  contributions: TContribution[]
+}