+++ src/types/crowdloanContribution.ts
@@ -1,0 +1,5 @@
+export type TContribution = {
+  kusamaAccount: string
+  contribution: string
+  memo: string
+}