+++ src/types/crowdloanFund.ts
@@ -1,0 +1,5 @@
+export type TCrowdloanFund = {
+  cap: string
+  raised: string
+  trieIndex: number
+}