+++ test/main.test.ts
@@ -1,13 +1,12 @@
-import { polkadot } from './dotSamaInstance'
+import { localTestnetRelay } from './dotSamaInstance'
 
 describe('test', () => {
   it('test', async () => {
-    const api = await polkadot.getApi()
+    const api = await localTestnetRelay.getApi()
     console.log(api.isConnected)
   })
 })
 
 afterAll(async () => {
-  await polkadot.disconnect()
-  await new Promise((resolve) => setTimeout(() => resolve(true), 10000))
+  await localTestnetRelay.disconnect()
 })