+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 ### 0.0.2 (2022-02-14)
+
+- Initial creation of Singleton
+- added method contributeToCrowdloan
+- added method getCrowdloanFunds